refactor: move code to own directories
