+++ src/index.js
@@ -1,4 +1,4 @@
-import printWelcomeMessage from "./welcome.js";
+import printWelcomeMessage from "./welcome/welcome.js";
 
 function main() {
   printWelcomeMessage();
